feat: add context menu support for all tabs
