+++ packages/bruno-app/src/components/RequestTabs/RequestTab/SpecialTab.js
@@ -1,8 +1,12 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import CloseTabIcon from './CloseTabIcon';
 import { IconVariable, IconSettings, IconRun, IconFolder, IconShieldLock } from '@tabler/icons';
+import { TabContextMenu } from 'components/TabContextMenu/index';
 
-const SpecialTab = ({ handleCloseClick, type, tabName, handleDoubleClick }) => {
+const SpecialTab = ({ handleCloseClick, type, tabName, handleDoubleClick, tabIndex }) => {
+  const dropdownTippyRef = useRef();
+  const onDropdownCreate = (ref) => (dropdownTippyRef.current = ref);
+
   const getTabInfo = (type, tabName) => {
     switch (type) {
       case 'collection-settings': {
@@ -27,7 +31,7 @@ const SpecialTab = ({ handleCloseClick, type, tabName, handleDoubleClick }) => {
             <IconShieldLock size={18} strokeWidth={1.5} className="text-yellow-600" />
             <span className="ml-1">Security</span>
           </>
-        )
+        );
       }
       case 'folder-settings': {
         return (
@@ -56,9 +60,26 @@ const SpecialTab = ({ handleCloseClick, type, tabName, handleDoubleClick }) => {
     }
   };
 
+  const handleRightClick = (_event) => {
+    const menuDropdown = dropdownTippyRef.current;
+    if (!menuDropdown) {
+      return;
+    }
+
+    if (menuDropdown.state.isShown) {
+      menuDropdown.hide();
+    } else {
+      menuDropdown.show();
+    }
+  };
+
   return (
     <>
-      <div className="flex items-center tab-label pl-2">{getTabInfo(type, tabName)}</div>
+      <div className="flex items-center tab-label pl-2" onContextMenu={handleRightClick}>
+        {getTabInfo(type, tabName)}
+        <TabContextMenu onDropdownCreate={onDropdownCreate} tabIndex={tabIndex} dropdownTippyRef={dropdownTippyRef} />
+      </div>
+
       <div className="flex px-2 close-icon-container" onClick={(e) => handleCloseClick(e)}>
         <CloseTabIcon />
       </div>
